+++ Sources/DcsInterface/DcsInterface.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "DcsSocket.h"
+#include "Utilities/DcsSocket.h"
 
 #include <map>
 #include <string>
