+++ Sources/DcsInterface/Utilities/DcsSocket.cpp
