+++ Sources/DcsInterface/Utilities/DcsSocket.h
