+++ Sources/DcsInterface/Utilities/Decimal.h
