+++ Sources/DcsInterface/Utilities/StringUtilities.cpp
