+++ Sources/DcsInterface/Utilities/StringUtilities.h
