+++ Sources/Windows/com.ctytler.dcs.sdPlugin.vcxproj
@@ -189,10 +189,10 @@
     <ClInclude Include="..\DcsInterface\DcsIdLookup.h" />
     <ClInclude Include="..\DcsInterface\DcsInterface.h" />
     <ClInclude Include="..\DcsInterface\DcsInterfaceParameters.h" />
-    <ClInclude Include="..\DcsInterface\DcsSocket.h" />
-    <ClInclude Include="..\DcsInterface\Decimal.h" />
+    <ClInclude Include="..\DcsInterface\Utilities\DcsSocket.h" />
+    <ClInclude Include="..\DcsInterface\Utilities\Decimal.h" />
     <ClInclude Include="..\DcsInterface\StreamdeckContext.h" />
-    <ClInclude Include="..\DcsInterface\StringUtilities.h" />
+    <ClInclude Include="..\DcsInterface\Utilities\StringUtilities.h" />
     <ClInclude Include="..\MyStreamDeckPlugin.h" />
     <ClInclude Include="pch.h" />
   </ItemGroup>
@@ -223,9 +223,9 @@
     </ClCompile>
     <ClCompile Include="..\DcsInterface\DcsIdLookup.cpp" />
     <ClCompile Include="..\DcsInterface\DcsInterface.cpp" />
-    <ClCompile Include="..\DcsInterface\DcsSocket.cpp" />
-    <ClCompile Include="..\DcsInterface\Decimal.cpp" />
-    <ClCompile Include="..\DcsInterface\StringUtilities.cpp" />
+    <ClCompile Include="..\DcsInterface\Utilities\DcsSocket.cpp" />
+    <ClCompile Include="..\DcsInterface\Utilities\Decimal.cpp" />
+    <ClCompile Include="..\DcsInterface\Utilities\StringUtilities.cpp" />
     <ClCompile Include="..\DcsInterface\StreamdeckContext.cpp" />
     <ClCompile Include="..\MyStreamDeckPlugin.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">/FI"pch.h" %(AdditionalOptions)</AdditionalOptions>
